[starts #161920708] create utilities and move decode token function into it
