+++ app/auth/v1/models.py
@@ -82,22 +82,6 @@ class User(Parent):
             return token
         except Exception as err:
             return str(err)
-        
-    @staticmethod
-    def decode_token(token):
-        """
-        method to decode the token generated during login
-        """
-        try:
-            #attempt to decode token using SECRET_KEY variable
-            payload = jwt.decode(token, current_app.config.get('SECRET_KEY'))
-            return payload['id']
-        except  jwt.ExpiredSignatureError:
-            # expired token returns an error string
-            return "Token expired. please login again to generate fresh token"
-        except jwt.InvalidTokenError:
-            #the token is not valid, throw error
-            return "Unworthy token. Please login to get fresh authorization"
 
     @classmethod
     def get_user_by_email(cls, email):
